mon: added ceph health check and removed aggregation of pool stats
Pool stats are no longer aggregated into an _all_ category - this can be done
above the collector. The net effect is that it allows a pools.* query to provide
all pools on a single chart - if the _all_ item was there this wouldn't be possible.

The mon_command interface is used for a "ceph health" command - which is
interpreted and pushed to graphite as a numeric value. This value can then be easily
plotted to see health over time.