added osd-state information
The mon collector now includes the status of osd's in it's output
(where 0 = up, 1 = down). To use this additional info, the at-a-glance
and backend-storage dashboards have been updated.

at-a-glance now includes a link from the OSD's 'down' value to the
backend storage dashboard

backend-storage dashboard includes an osd down table - which basically
shows osd's with a status of > 0