at-a-glance: fix for health panel, and colour match with status panel
The singlestat panel was using value map, but singlestat appears to be interpolating
the health value which results in the value map not being used and a number appearing
on the dashboard. This update uses a range map to handle/workaround this nuance.

In addition, cosmetic changes to the health, disk and latency panels - making their warning
colour state match the status panel warning colour for consistency