cephmetrics: Use a default value type for graphite, and assign default logging
Before this change, if a variable was defined in a class but NOT defined in it's attributes
the collector would fail. With this change a default of gauge is assigned.

In addition, a default logging level is set for all the collectors, if not specified by LogLevel
in the collectd.conf plugin