base: use basic logging facility
A separate logging class was used, but with this in place the
filename,funName parameters always showed the method in the logging
class not the file/function of the calling module.

This base class now just grabs the cephmetrics named logger setup in the
collectd parent module, which is inherited by the OSDs/RGW and Mon
classes.