mon: calculation of osd host count changed, and additional per osd info added
Instead of using 'ceph osd tree' to determine the unique hosts in the
cluster, 'ceph osd dump' is used. The dump provides the IP address of
the owner of the OSD, so this can be used with a python set to derive
the unique IP's - which infers the number of physical hosts in the
cluster.

'osd tree' is problematic when crushmap is used for things like pools,
since the hostnames in the map are logical to crush, not physical names
within the cluster.