osd-node-detail: fixes to allow the dashboard to be used directly
Initially the intention was this dashboard would be opened by navigating
to it by drilling through higher level dashboards. However, it is being
used directly so the osd_servers and disks template variable needed to
be exposed to make the charts usable. Without these variables selectable
the charts are unreadable.

In addition an OSD count and a breakdown of the disk -> OSD id is also
provided, also tied to the same templating variables