mon: ceph health checks exposed through the event url
The parent Ceph class has an event_url attribute that is used by Mon to
send text based state to a given endpoint. Ceph health states are compared
at each get_stats iteration, and when the overall health changes or
the context of the health event changes the health text is sent.

In addition to the event information, the mon now exposes the ceph
version (ceph-mon --version)