osd: key bluestore metrics added and updated to soft-fail on missing osd socket
Bluestore provides a multitude of perf metrics, of which a subset have
been chosen that serve as health/performance indicators for bluestore
osd performance.
Also, since a failed OSD will not have an OSD socket the collector now just
treats the missing socket as an error, reports it up to the Ceph class through
the instances error boolean, so it can be exposed to the caller - instead
of simply raising an exception.