at-a-glance db: host panels for OSD and RGW hosts switched to status-panel
Status panel now being used so the panel can show total/up and down
metrics for the host types. This up/down state is purely based on whether
collectd is getting and sending data. The count for total is based on the
number of metrics seen in the graphite backend.