ansible: Do not enable rhsm repos
We are shipping as a regular product and as such, we cannot enable
additional repos via rhsm. The customers might even not have these repos
installed (especially the storage console repo might not be available to
them).

If we need any of the packages from these repos, we need to cross-ship
them in our product (as we already do downstream).

Signed-off-by: Boris Ranto <branto@redhat.com>