prometheus: Simplify dashboards
This commit removes the old no longer used lines from the prometheus
dashboards. The lines being removed are used by graphite (or influxdb).
The grafana web UI does not overwrite these and keeps them stored in its
db when you are doing changes to the grafana dashboards with web UI.

Signed-off-by: Boris Ranto <branto@redhat.com>