Silence the selinux denial on loading a module
If the cfg80211 module is not loaded and node_exporter will try to
detect a presence of wireless network devices, the module gets
auto-loaded. This results in an SELinux denial since the container is
trying to load a module. Working around this by making sure the module
is always loaded before the container is started.

Signed-off-by: Boris Ranto <branto@redhat.com>